requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
greenlet==2.0.2
idna==3.2
loguru==0.7.0
pyTelegramBotAPI==4.6.0
python-dotenv==1.0.0
requests==2.26.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
urllib3==1.26.6
